explorer/requirements.txt
base58==1.0.3
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colander==1.7.0
ecdsa==0.13.2
Flask==1.1.1
Flask-CDN==1.5.3
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.1
libeospy==2.0.1
MarkupSafe==1.1.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
translationstring==1.3
urllib3==1.25.3
Werkzeug==0.15.5
